refactor and use vcpkg
