+++ .gitignore
@@ -1,6 +1,9 @@
 # Prerequisites
 *.d
 
+# CLion
+.idea
+
 # Compiled Object files
 *.slo
 *.lo
