--- .idea/Cosmo_Ai.iml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="FacetManager">
-    <facet type="Python" name="Python facet">
-      <configuration sdkName="Python 3.8 (Cosmo_Ai)" />
-    </facet>
-  </component>
-  <component name="NewModuleRootManager">
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="library" name="Python 3.8 (Cosmo_Ai) interpreter library" level="application" />
-  </component>
-</module>
