--- .idea/Cosmo_Ai2.iml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module classpath="CMake" type="CPP_MODULE" version="4">
-  <component name="FacetManager">
-    <facet type="Python" name="Python facet">
-      <configuration sdkName="Python 3.8 (Cosmo_Ai)" />
-    </facet>
-  </component>
-</module>
