--- .idea/modules.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/CosmoAi.iml" filepath="$PROJECT_DIR$/.idea/CosmoAi.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/CosmoAi2.iml" filepath="$PROJECT_DIR$/.idea/CosmoAi2.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/Cosmo_Ai.iml" filepath="$PROJECT_DIR$/.idea/Cosmo_Ai.iml" />
-    </modules>
-  </component>
-</project>
