+++ src/makeimage/makeimage.cpp
