+++ src/simlib/EllipsoidSource.cpp
@@ -1,6 +1,7 @@
 /* (C) 2022: Hans Georg Schaathun <georg@schaathun.net> */
 
-#include "Simulator.h"
+#include "cosmosim/Simulator.h"
+
 #include <thread>
 
 EllipsoidSource::EllipsoidSource( int sz, double sig1, double sig2, double thet ) :
