+++ src/simlib/LensModel.cpp
@@ -3,7 +3,8 @@
  * Einar Leite Austnes, and Simon Nedreberg Runde
  */
 
-#include "Simulator.h"
+#include "cosmosim/Simulator.h"
+
 #include <symengine/expression.h>
 #include <symengine/lambda_double.h>
 #include <thread>
