+++ vcpkg.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://raw.githubusercontent.com/microsoft/vcpkg/master/scripts/vcpkg.schema.json",
+  "name": "cosmosim",
+  "dependencies": [
+    "symengine",
+    "opencv",
+    "pybind11"
+  ]
+}
